requirements.txt
playwright==1.40.0
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
